Fixed no devices found

+++ getports.sh
@@ -75,9 +75,11 @@ doPort() {
 ############
 
 doTar() {
-  echo -e "\nAdding files to tarball:"
-  find "$HOMEPATH" -name "*.device" -print0 | tar -cvzf "$HOMEPATH/$TARBALL" --null -T -
-  find "$HOMEPATH" -name "*.device" -type f -exec rm '{}' \;
+  if ls "$HOMEPATH"/*.devices 1> /dev/null 2>&1; then
+    echo -e "\nAdding files to tarball:"
+    find "$HOMEPATH" -name "*.device" -print0 | tar -cvzf "$HOMEPATH/$TARBALL" --null -T -
+    find "$HOMEPATH" -name "*.device" -type f -exec rm '{}' \;
+  fi
 }
 
 ############
@@ -97,7 +99,11 @@ main() {
   checkroot
   doPort
   [ "$devNum" -gt 0 ] && doTar
-  "$DOFTP" && doFTP || echo -e "\nPlease email $HOMEPATH/$TARBALL to Lee@Bussy.org."
+  if [ -d $HOMEPATH/$TARBALL ]; then
+    "$DOFTP" && doFTP || echo -e "\nPlease email $HOMEPATH/$TARBALL to Lee@Bussy.org."
+  else
+    echo -e "\nNo devices found."
+  fi
 }
 
 ############
